crawler/requirements.txt
attrs==16.3.0 # Updated from 16.1.0
cffi==1.9.1 # Updated from 1.7.0
ConcurrentLogHandler==0.9.1
cryptography==1.8.1 # Updated from 1.5
cssselect==1.0.1 # Updated from 0.9.2
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0 # Updated from 0.15.2
idna==2.5 # Updated from 2.1
ipaddress==1.0.18 # Updated from 1.0.16
kafka-python==1.3.3 # Updated from 1.3.2
kazoo==2.2.1
lxml==3.7.3 # Updated from 3.6.4
mock==2.0.0
nose==1.3.7
parsel==1.1.0 # Updated from 1.0.3
pbr==2.0.0 # Updated from 1.10.0
pyasn1==0.2.3 # Updated from 0.1.9
pyasn1-modules==0.0.8
pycparser==2.17 # Updated from 2.14
PyDispatcher==2.0.5
pyOpenSSL==16.2.0 # Updated from 16.1.0
python-json-logger==0.1.7 # Updated from 0.1.5
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
requests==2.13.0 # Updated from 2.11.1
requests-file==1.4.1 # Updated from 1.4
retrying==1.3.3
Scrapy==1.3.3
scutils==1.2.0
service-identity==16.0.0
six==1.10.0
testfixtures==4.13.5 # Updated from 4.10.0
tldextract==2.0.2 # Updated from 2.0.1
Twisted==17.1.0 # Updated from 16.4.0
ujson==1.35
w3lib==1.17.0 # Updated from 1.16.0
zope.interface==4.3.3 # Updated from 4.2.0
# Generated with piprot 0.9.7
